Lorem

+++ Part_III/Spreadsheet.cpp
@@ -1,22 +1,22 @@
 #include "Spreadsheet.h"
 #include "SpreadsheetImpl.h"
 
-void Spreadsheet::setCellAt(size_t x, size_t y, const SpreadsheetCell& inCell)
+void Spreadsheet::setCellAt(int x, int y, const SpreadsheetCell& inCell)
 {
 	mImpl->setCellAt(x, y, inCell);
 }
 
-SpreadsheetCell& Spreadsheet::getCellAt(size_t x, size_t y)
+SpreadsheetCell& Spreadsheet::getCellAt(int x, int y)
 {
 	return mImpl->getCellAt(x, y);
 }
 
-size_t Spreadsheet::getId() const
+int Spreadsheet::getId() const
 {
 	return mImpl->getId();
 }
 
-Spreadsheet::Spreadsheet(const SpreadsheetApplication &theApp, size_t inWidth, size_t inHeight)
+Spreadsheet::Spreadsheet(const SpreadsheetApplication &theApp, int inWidth, int inHeight)
 {
 	mImpl = std::make_shared<SpreadsheetImpl>(theApp, inWidth, inHeight);
 }
