+++ Part_III/Spreadsheet.h
@@ -11,18 +11,18 @@ class Spreadsheet
 {
 public:
 	Spreadsheet(const SpreadsheetApplication& theApp,
-		size_t inWidth = kMaxWidth, size_t inHeight = kMaxHeight);
+		int inWidth = kMaxWidth, int inHeight = kMaxHeight);
 	Spreadsheet(const Spreadsheet& src);
 
 	Spreadsheet& operator=(const Spreadsheet& rhs);
 	
-	void setCellAt(size_t x, size_t y, const SpreadsheetCell& inCell);
-	SpreadsheetCell& getCellAt(size_t x, size_t y);
+	void setCellAt(int x, int y, const SpreadsheetCell& inCell);
+	SpreadsheetCell& getCellAt(int x, int y);
 	
-	size_t getId() const;
+	int getId() const;
 
-	static const size_t kMaxHeight = 100;
-	static const size_t kMaxWidth = 100;
+	static const int kMaxHeight = 100;
+	static const int kMaxWidth = 100;
 
 private:
 	std::shared_ptr<SpreadsheetImpl> mImpl;
